Make the source image part of the filter environment
Added a user_data field to the _envir structure to store a pointer to
additional data that can be used by the callback functions.
Added a SourceImageData structure to provide the GetPixel method with
access the source image data without requiring global variables.

Updated the CreateEnvironmentData method to use the SourceImageData
structure and set the user_data field in the _envir structure.
Updated the FreeEnvironmentData method to free the user_data field in
the _envir structure.

Removed the Pixeldata, Stride and PixelSize global variables.
Removed the SetupBitmap and DestroyBitmap methods.